test: colocate runner regression tests with their implementations
The coverage-ci threshold test and the run-test-file environment test
lived under tests/integration/test-runner/, against the colocation rule
and outside the test:scripts discovery list. Fold them into the existing
scripts/test/coverage-ci.test.ts and scripts/test/run-test-file.test.ts,
which test:scripts already runs.
