+++ scripts/test/coverage-ci.test.ts
@@ -1,4 +1,4 @@
-import { fromFileUrl } from "#std/path";
+import { fromFileUrl, join } from "#std/path";
 import { assert, assertEquals } from "#veryfront/testing/assert.ts";
 import { describe, it } from "#veryfront/testing/bdd.ts";
 import {
@@ -68,6 +68,39 @@ describe("coverage CI command", () => {
         coverageExcludePatterns().some((pattern) => pattern.test(path)),
         `${path} must be excluded from coverage`,
       );
+    }
+  });
+
+  it("does not treat a separate threshold value as an LCOV path", async () => {
+    const repoRoot = fromFileUrl(new URL("../../", import.meta.url));
+    const tempDir = await Deno.makeTempDir();
+    try {
+      const output = await new Deno.Command(Deno.execPath(), {
+        args: [
+          "run",
+          `--config=${join(repoRoot, "scripts/test.deno.json")}`,
+          "--no-npm",
+          "--allow-read",
+          "--allow-write",
+          join(repoRoot, "scripts/test/coverage-ci.ts"),
+          "merge",
+          "--threshold",
+          "85",
+          "missing-lcov",
+        ],
+        cwd: tempDir,
+        stdout: "piped",
+        stderr: "piped",
+      }).output();
+      const stderr = new TextDecoder().decode(output.stderr);
+
+      assertEquals(output.success, false);
+      assert(
+        stderr.includes("missing-lcov"),
+        `expected the positional LCOV path in the failure, got: ${stderr}`,
+      );
+    } finally {
+      await Deno.remove(tempDir, { recursive: true });
     }
   });
